requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
py-autoclean>=1.1.0
google-generativeai>=0.4.0
openai>=1.0.0
plotly>=5.18.0
pydantic>=2.6.0